+++ skill/mdoc/mdoc_core/__init__.py
@@ -1,3 +1,3 @@
 """mdoc core package."""
 
-VERSION = "1.5.6"
+VERSION = "1.5.7"